Allowed transactions to be edited. Date selection during edits

+++ YouthApp/Controllers/TransactionItemsController.cs
@@ -18,7 +18,7 @@ namespace YouthApp.Controllers
         public async Task<IEnumerable> List() => await new ApplicationDbContext(dco).TransactionItems.OrderBy(x => x.TransactionItem).Take(20).ToListAsync();
 
         [HttpGet]
-        public async Task<IActionResult> Find(int id)
+        public async Task<IActionResult> Find(short id)
         {
             var tran = await new ApplicationDbContext(dco).TransactionItems.FindAsync(id);
             if (tran == null)
